Triple and Quad moved to util package
